Factor button out and use in sidebar
