+++ browser/src/UI/components/SidebarEmptyPaneView.tsx
@@ -20,24 +20,8 @@ const Wrapper = styled.div`
     justify-content: center;
 `
 
-import { boxShadow, withProps } from "./common"
-
-const ButtonWrapper = styled.button`
-    background-color: ${props => props.theme.background};
-    color: ${props => props.theme.foreground};
-    padding: 1em;
-    border: 2px solid transparent;
-    width: 100%;
-    outline: none;
-    cursor: pointer;
-    transition: all 0.1s ease-in;
-
-    pointer-events: all;
-
-    &:hover {
-        ${boxShadow} transform: translateY(-1px);
-    }
-`
+import { SidebarButton } from "./SidebarButton"
+import { VimNavigator } from "./VimNavigator"
 
 const Description = styled.div`
     margin: 32px;
@@ -45,48 +29,10 @@ const Description = styled.div`
     text-align: center;
 `
 
-export interface ButtonContainerProps {
-    selected: boolean
-}
-
-const ButtonContainer = withProps<ButtonContainerProps>(styled.div)`
-    padding-left: 32px;
-    padding-right: 32px;
-
-    transition: all 0.1s ease-in;
-
-    background-color: ${props => (props.selected ? "rgba(0, 0, 0, 0.1)" : "transparent")};
-    border-left: 2px solid ${props =>
-        props.selected ? props.theme["highlight.mode.normal.background"] : "transparent"};
-`
-
-import { Sneakable } from "./Sneakable"
-import { VimNavigator } from "./VimNavigator"
-
-export interface IOniButtonProps {
-    focused: boolean
-    text: string
-    onClick: () => void
-}
-
-export class OniButton extends React.PureComponent<IOniButtonProps, {}> {
-    public render(): JSX.Element {
-        return (
-            <ButtonContainer selected={this.props.focused}>
-                <Sneakable callback={this.props.onClick}>
-                    <ButtonWrapper onClick={this.props.onClick}>
-                        <span>{this.props.text}</span>
-                    </ButtonWrapper>
-                </Sneakable>
-            </ButtonContainer>
-        )
-    }
-}
-
 export class SidebarEmptyPaneView extends React.PureComponent<ISidebarEmptyPaneViewProps, {}> {
     public render(): JSX.Element {
         const button = this.props.actionButtonText ? (
-            <OniButton
+            <SidebarButton
                 focused={this.props.active}
                 text={this.props.actionButtonText}
                 onClick={() => this.props.onClickButton && this.props.onClickButton()}
